refactor(events): app direct related events in separate file
